Cohen's Effect Size

Computes the Cohen's d and Hedges'g effect size statistics.

cohens_d(x, y = NULL, pooled = TRUE, correct = FALSE, conf.level = NA,
  na.rm = FALSE, group = NULL, ...)

# S3 method for default
cohens_d(x, y = NULL, pooled = TRUE, correct = FALSE,
  conf.level = NA, na.rm = FALSE, group = NULL, ...)

Arguments

x
a (non-empty) numeric vector of data values.

y
a (non-empty) numeric vector of data values.

pooled
(logical) Flag indicating whether compute pooled standard deviation or the whole sample standard deviation. Default is TRUE.

correct
(logical) Flag indicating whether to apply the Hedges correction. (Default: FALSE)

conf.level
(numeric|NA) confidence level of the interval. Set this to NA, if no confidence intervals should be calculated. (This is the default)

na.rm
(logical) Should missing values be removed? Defaults to FALSE.

Value

A numeric vector with 3 elements:

dthe effect size d

lwr.cilower bound of the confidence interval

upr.ciupper bound of the confidence interval

Origin of the code

The main part of code and documentation of this function is imported from: Andri Signorell et al. (2017). DescTools: Tools for descriptive statistics. R package version 0.99.22 (2017-09-14).

The modifications were made by Vilmantas Gegzna. The logics of statistical computations were not changed.

References

Cohen, J. (1988) Statistical power analysis for the behavioral sciences (2nd ed.) Academic Press, New York.

Hedges, L. V. & Olkin, I. (1985) Statistical methods for meta-analysis Academic Press, Orlando, FL

Smithson, M.J. (2003) Confidence Intervals, Quantitative Applications in the Social Sciences Series, No. 140. Thousand Oaks, CA: Sage. pp. 39-41

Cohen, J. (1992). A power primer. Psychological Bulletin, 112, 155-159.

See also

mean, var

Examples

data("d.pizza", package = "DescTools")
library(biostat)

x <- d.pizza$price[d.pizza$driver == "Carter"]
y <- d.pizza$price[d.pizza$driver == "Miller"]

cohens_d(x, y, conf.level = 0.95, na.rm = TRUE)

#>             d lwr.ci.lwr.ci  upr.ci.lprob 
#>    -0.2122779    -1.9835187     0.9750000 
#> attr(,"magnitude")
#> [1] "small"