Games-Howell and Tukey post-hoc tests

Either Games-Howell or Tukey post-hoc tests for one-way analysis of variance (ANOVA).

posthocTGH(y, x, method = c("games-howell", "tukey"), conf.level = 0.95,
  digits = 2, p.adjust = "holm", formatPvalue = TRUE)

# S3 method for posthocTGH
print(x, digits = x$input$digits, ...)

Arguments

y
y (numeric) a numeric vector.

x
x (factor) a vector that either is a factor or can be converted into one.

method
("games-howell"|"tukey") Which post-hoc tests to conduct. Valid values are "tukey" and "games-howell".

conf.level
Confidence level of the confidence intervals.

digits
The number of digits to show in the output.

p.adjust
Any valid p.adjust method.

formatPvalue
Whether to format the p values according to APA standards (i.e. replace all values lower than .001 with '<.001'). This only applies to the printing of the object, not to the way the p values are stored in the object.

Value

A list of three elements:

input

List with input arguments

intermediate

List of intermediate objects.

output

List with two objects 'tukey' and 'games.howell', containing the outcomes for the respective post-hoc tests.

Details

This function is omported from package userfriendlyscience (posthocTGH)

Note

This function is based on a file that was once hosted at http://www.psych.yorku.ca/cribbie/6130/games_howell.R, but has been removed since. It was then adjusted for implementation in the userfriendlyscience package. Jeffrey Baggett needed the confidence intervals, and so emailed them, after which his updated functions was used. In the meantime, it appears Aaron Schlegel (https://rpubs.com/aaronsc32) independently developed a version with confidence intervals and posted it on RPubs at https://rpubs.com/aaronsc32/games-howell-test.

See also

http://www.gcf.dkf.unibe.ch/BCB/files/BCB_10Jan12_Alexander.pdf

https://rpubs.com/aaronsc32/games-howell-test

http://aoki2.si.gunma-u.ac.jp/R/src/tukey.R

https://stats.stackexchange.com/questions/83941/games-howell-post-hoc-test-in-r

Examples

library(BioStat)

### Compute post-hoc statistics using the Games-Howell method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet)

#>     n means variances
#> 1 220   103      3210
#> 2 120   123      5128
#> 3 120   143      7489
#> 4 118   135      4737
#> 
#>     diff  ci.lo ci.hi    t  df     p p.adjusted
#> 2-1 20.0   0.36    40 2.64 201  0.04       0.18
#> 3-1 40.3  17.54    63 4.59 176 <0.01      <0.01
#> 4-1 32.6  13.45    52 4.41 203 <0.01      <0.01
#> 3-2 20.3  -6.20    47 1.98 230  0.20       0.59
#> 4-2 12.6 -10.91    36 1.39 236  0.51       1.00
#> 4-3 -7.7 -33.90    19 0.76 226  0.87       1.00

### Compute post-hoc statistics using the Tukey method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet, method="tukey");

#>     n means variances
#> 1 220   103      3210
#> 2 120   123      5128
#> 3 120   143      7489
#> 4 118   135      4737
#> 
#>     diff ci.lo ci.hi    t  df     p p.adjusted
#> 2-1 20.0  -0.3    40 2.54 574  0.06       0.22
#> 3-1 40.3  20.0    61 5.12 574 <0.01      <0.01
#> 4-1 32.6  12.2    53 4.12 574 <0.01      <0.01
#> 3-2 20.3  -2.7    43 2.27 574  0.11       0.32
#> 4-2 12.6 -10.5    36 1.41 574  0.50       0.99
#> 4-3 -7.7 -30.8    15 0.86 574  0.83       0.99