! Compute data for a qq-plot
Source: R/qq_data.R

qq_data(
  y,
  data = NULL,
  distribution = "norm",
  ...,
  line = c("quartiles", "robust", "int=0,slope=1", "0,1", "none"),
  envelope = 0.95,
  method = c("mle-normal", "trimmed-normal", "moment-normal", "any"),
  labels = NULL,
  groups = NULL,
  sep = " | "
)

# S3 method for qqdata
coef(object, ...)

# S3 method for qqdata
plot(x, ..., use_colors = FALSE, scales = "free")

Arguments

y
(formula|numeric|character)
Either a formula, a numeric vector or a name of a column in data.

If y is a formula (e.g. variable ~ factor), left-hand side provides the name of a variable to be tested. In the right-hand side there are names of factor variables to be used to create subsets. If the left-hand side is empty (e.g. ~ factor), right-hand side is treated as variable name to test.

data
(data frame|NULL)
Either a data frame that contains the variables mentioned in y or NULL (if the variables are in the function's environment).

...
Parameters to be passed to function, selected in distribution. In print method, further parameters to function print.

line
(string) A parameter, that controls how reference line is drawn. Options:

"0,1" or "int=0,slope=1" to plot a line with intercept = 0 and slope = 1;

"quartiles" to pass a line through the quartile-pairs;

"robust" for a robust-regression line; the latter uses the rlm function from the MASS package;

option "none" is not implemented yet.

envelope
(numeric | FALSE) confidence level for point-wise confidence envelope.

method
(string: "trimmed-normal", "normal", "any"). A method, that controls how estimates of parameters for distribution are computed. Options:

groups(NULL|factor|character)
An alternative way to provide groups. If y is a numeric vector, groups must be a factor (or NULL). If y is a sting, groups must also be a string (or NULL).

sep(not used yet).

objectA qqdata object.

xA qqdata object.

use_colors(logical) use colors for multiple groups

scales("free"|"free_x"|"free_y"|"fixed") a parmeter to be passed to facet_wrap.