Do numerical summaries by groups

Do numerical summaries by groups with formaula interface. Missing values are automatically removed.

do_summary(y, data = NULL, stat = c("n", "missing", "mean", "sd", "var",
  "min", "Q1", "Md", "Q3", "max", "IQR", "mad", "skewness", "kurtosis"))

# S3 method for num_summaries
print(x, ..., digits = NA, format = "f")

Arguments

y
formula with variable names to summarize. See more in examples.

data
data set

stat
(character) Descriptive statistics to compute. Currently supported statistics:

"n" - number of non-missing observations,

"missing" - number of missing observations,,

"mean" - arithmetic mean,

"sd" - standard deviation,

"var" - variance,

"min" - minimum value,

"Q1" - 1-st quartile,

"Md" - median,

"Q3" - 3-rd quartile,

"max" - maximum value,

"IQR" - interquartile range,

"mad" - median absolute deviation from median (more details mad),

"skewness" - skewness,

"kurtosis" - excess kurtosis.

x
object to print

...
further arguments to methods.

digits
(numeric or NA) a desired number of digits after the decimal point (if format = "f") or a number of significant digits (format = "g", = "e" or = "fg"). Either one integer (to set the same formatting for all columns) or a vector of inegers (to set formatting for each column separately). Use NA to leave column unformatted.

format
(character) "f", "g", "e", "fg". Either one value or a vector of values for each column. Each value will be passed to fun separately. "f" gives numbers in the usual xxx.xxx format; "e" and "E" give n.ddde+nn or n.dddE+nn (scientific format); "g" and "G" put number into scientific format only if it saves space to do so. "fg" uses fixed format as "f", but digits as the minimum number of significant digits. This can lead to quite long result strings

Value

Data frame with summary satatistics.

Examples

library(biostat)
data(cabbages, package = "MASS")

do_summary(~VitC, data = cabbages) %>%
    print(digits = 2)

#>   summary_of  n missing  mean    sd    var   min    Q1    Md    Q3   max   IQR
#> 1       VitC 60       0 57.95 10.12 102.39 41.00 50.75 56.00 66.25 84.00 15.50
#>     mad skewness kurtosis
#> 1 10.38     0.32    -0.68

do_summary(VitC ~ Cult, data = cabbages) %>%
    print(digits = 2)

#>   summary_of Cult  n missing  mean   sd   var   min    Q1    Md    Q3   max
#> 1       VitC  c39 30       0 51.50 7.12 50.74 41.00 46.00 51.00 54.75 68.00
#> 2       VitC  c52 30       0 64.40 8.46 71.49 47.00 58.00 64.50 70.75 84.00
#>     IQR  mad skewness kurtosis
#> 1  8.75 5.93     0.58    -0.26
#> 2 12.75 9.64     0.12    -0.62

do_summary(VitC ~ Cult + Date, data = cabbages, stat = "mean") %>%
    print(digits = 2)

#>   summary_of Cult Date  mean
#> 1       VitC  c39  d16 50.30
#> 2       VitC  c39  d20 49.40
#> 3       VitC  c39  d21 54.80
#> 4       VitC  c52  d16 62.50
#> 5       VitC  c52  d20 58.90
#> 6       VitC  c52  d21 71.80

do_summary(HeadWt + VitC ~ Cult + Date,
           data = cabbages,
           stat = c("n", "mean", "sd")) %>%
    print(digits = 1)

#>    summary_of Cult Date  n mean  sd
#> 1      HeadWt  c39  d16 10  3.2 1.0
#> 2      HeadWt  c39  d20 10  2.8 0.3
#> 3      HeadWt  c39  d21 10  2.7 1.0
#> 4      HeadWt  c52  d16 10  2.3 0.4
#> 5      HeadWt  c52  d20 10  3.1 0.8
#> 6      HeadWt  c52  d21 10  1.5 0.2
#> 7        VitC  c39  d16 10 50.3 4.3
#> 8        VitC  c39  d20 10 49.4 8.3
#> 9        VitC  c39  d21 10 54.8 7.6
#> 10       VitC  c52  d16 10 62.5 5.8
#> 11       VitC  c52  d20 10 58.9 7.7
#> 12       VitC  c52  d21 10 71.8 6.2